@@ -1,6 +1,6 @@
 {
   "name": "new.expensify",
-  "version": "1.3.56-0",
+  "version": "1.3.56-18",
   "author": "Expensify, Inc.",
   "homepage": "https://new.expensify.com",
   "description": "New Expensify is the next generation of Expensify: a reimagination of payments based atop a foundation of chat.",
@@ -147,7 +147,7 @@
     "react-native-web-linear-gradient": "^1.1.2",
     "react-native-web-lottie": "^1.4.4",
     "react-native-webview": "^11.17.2",
-    "react-native-x-maps": "1.0.6",
+    "react-native-x-maps": "1.0.10",
     "react-pdf": "^6.2.2",
     "react-plaid-link": "3.3.2",
     "react-web-config": "^1.0.0",
@@ -234,7 +234,6 @@
     "eslint-plugin-react-native-a11y": "^3.3.0",
     "eslint-plugin-storybook": "^0.5.13",
     "eslint-plugin-you-dont-need-lodash-underscore": "^6.12.0",
-    "flipper-plugin-bridgespy-client": "^0.1.9",
     "html-webpack-plugin": "^5.5.0",
     "jest": "29.4.1",
     "jest-circus": "29.4.1",
@@ -257,7 +256,7 @@
     "style-loader": "^2.0.0",
     "time-analytics-webpack-plugin": "^0.1.17",
     "type-fest": "^3.12.0",
-    "typescript": "^4.8.4",
+    "typescript": "^5.1.6",
     "wait-port": "^0.2.9",
     "webpack": "^5.76.0",
     "webpack-bundle-analyzer": "^4.5.0",
