@@ -15,12 +15,17 @@ const emojiNameTable = _.reduce(
     {},
 );
 
-const emojiCodeTable = _.reduce(
+const emojiCodeTableWithSkinTones = _.reduce(
     emojis,
     (prev, cur) => {
         const newValue = prev;
         if (!cur.header) {
             newValue[cur.code] = cur;
+        }
+        if (cur.types) {
+            cur.types.forEach((type) => {
+                newValue[type] = cur;
+            });
         }
         return newValue;
     },
@@ -32,5 +37,5 @@ const localeEmojis = {
     es: esEmojis,
 };
 
-export {emojiNameTable, emojiCodeTable, localeEmojis};
+export {emojiNameTable, emojiCodeTableWithSkinTones, localeEmojis};
 export {skinTones, categoryFrequentlyUsed, default} from './common';
