@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import CONST from '../../CONST';
 import useLocalize from '../../hooks/useLocalize';
 import HeaderWithBackButton from '../HeaderWithBackButton';
-import SelectionListRadio from '../SelectionListRadio';
+import SelectionList from '../SelectionList';
 import Modal from '../Modal';
 import ScreenWrapper from '../ScreenWrapper';
 import styles from '../../styles/styles';
@@ -73,7 +73,7 @@ function CountrySelectorModal({currentCountry, isVisible, onClose, onCountrySele
                     title={translate('common.country')}
                     onBackButtonPress={onClose}
                 />
-                <SelectionListRadio
+                <SelectionList
                     headerMessage={headerMessage}
                     textInputLabel={translate('common.country')}
                     textInputPlaceholder={translate('countrySelectorModal.placeholderText')}
@@ -81,8 +81,6 @@ function CountrySelectorModal({currentCountry, isVisible, onClose, onCountrySele
                     sections={[{data: searchResults, indexOffset: 0}]}
                     onSelectRow={onCountrySelected}
                     onChangeText={setSearchValue}
-                    shouldFocusOnSelectRow
-                    shouldHaveOptionSeparator
                     shouldDelayFocus
                     initiallyFocusedOptionKey={currentCountry}
                 />
