@@ -1,9 +1,9 @@
 import React, {forwardRef} from 'react';
 import {Keyboard} from 'react-native';
-import BaseSelectionListRadio from './BaseSelectionListRadio';
+import BaseSelectionList from './BaseSelectionList';
 
-const SelectionListRadio = forwardRef((props, ref) => (
-    <BaseSelectionListRadio
+const SelectionList = forwardRef((props, ref) => (
+    <BaseSelectionList
         // eslint-disable-next-line react/jsx-props-no-spreading
         {...props}
         ref={ref}
@@ -12,6 +12,6 @@ const SelectionListRadio = forwardRef((props, ref) => (
     />
 ));
 
-SelectionListRadio.displayName = 'SelectionListRadio';
+SelectionList.displayName = 'SelectionList';
 
-export default SelectionListRadio;
+export default SelectionList;
