@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import CONST from '../../CONST';
 import Modal from '../Modal';
 import HeaderWithBackButton from '../HeaderWithBackButton';
-import SelectionListRadio from '../SelectionListRadio';
+import SelectionList from '../SelectionList';
 import useLocalize from '../../hooks/useLocalize';
 import ScreenWrapper from '../ScreenWrapper';
 import styles from '../../styles/styles';
@@ -78,7 +78,7 @@ function StateSelectorModal({currentState, isVisible, onClose, onStateSelected, 
                     shouldShowBackButton
                     onBackButtonPress={onClose}
                 />
-                <SelectionListRadio
+                <SelectionList
                     headerMessage={headerMessage}
                     textInputLabel={label || translate('common.state')}
                     textInputPlaceholder={translate('stateSelectorModal.placeholderText')}
@@ -86,8 +86,6 @@ function StateSelectorModal({currentState, isVisible, onClose, onStateSelected, 
                     sections={[{data: searchResults, indexOffset: 0}]}
                     onSelectRow={onStateSelected}
                     onChangeText={setSearchValue}
-                    shouldFocusOnSelectRow
-                    shouldHaveOptionSeparator
                     shouldDelayFocus
                     initiallyFocusedOptionKey={currentState}
                 />
