@@ -2,16 +2,24 @@ import _ from 'underscore';
 import React from 'react';
 
 const composerRef = React.createRef();
+// There are two types of composer: general composer (edit composer) and main composer.
+// The general composer callback will take priority if it exists.
 let focusCallback = null;
+let mainComposerFocusCallback = null;
 
 /**
  * Register a callback to be called when focus is requested.
  * Typical uses of this would be call the focus on the ReportActionComposer.
  *
  * @param {Function} callback callback to register
+ * @param {Boolean} isMainComposer
  */
-function onComposerFocus(callback) {
-    focusCallback = callback;
+function onComposerFocus(callback, isMainComposer = false) {
+    if (isMainComposer) {
+        mainComposerFocusCallback = callback;
+    } else {
+        focusCallback = callback;
+    }
 }
 
 /**
@@ -20,6 +28,11 @@ function onComposerFocus(callback) {
  */
 function focus() {
     if (!_.isFunction(focusCallback)) {
+        if (!_.isFunction(mainComposerFocusCallback)) {
+            return;
+        }
+
+        mainComposerFocusCallback();
         return;
     }
 
@@ -29,9 +42,14 @@ function focus() {
 /**
  * Clear the registered focus callback
  *
+ * @param {Boolean} isMainComposer
  */
-function clear() {
-    focusCallback = null;
+function clear(isMainComposer = false) {
+    if (isMainComposer) {
+        mainComposerFocusCallback = null;
+    } else {
+        focusCallback = null;
+    }
 }
 
 /**
