@@ -426,6 +426,8 @@ function getMoneyRequestInformation(report, participant, comment, amount, curren
         optimisticTransaction.transactionID,
         '',
         iouReport.reportID,
+        false,
+        false,
         receiptObject,
     );
 
@@ -1030,7 +1032,7 @@ function editMoneyRequest(transactionID, transactionThreadReportID, transactionC
     ];
 
     // STEP 6: Call the API endpoint
-    const {created, amount, currency, comment} = ReportUtils.getTransactionDetails(updatedTransaction);
+    const {created, amount, currency, comment, merchant} = ReportUtils.getTransactionDetails(updatedTransaction);
     API.write(
         'EditMoneyRequest',
         {
@@ -1040,6 +1042,7 @@ function editMoneyRequest(transactionID, transactionThreadReportID, transactionC
             amount,
             currency,
             comment,
+            merchant,
         },
         {optimisticData, successData, failureData},
     );
