@@ -3,6 +3,7 @@ import Onyx from 'react-native-onyx';
 import lodashGet from 'lodash/get';
 import ONYXKEYS from '../../ONYXKEYS';
 import * as CollectionUtils from '../CollectionUtils';
+import * as API from '../API';
 
 const allTransactions = {};
 Onyx.connect({
@@ -47,6 +48,15 @@ function addStop(transactionID) {
                 [`waypoint${newLastIndex}`]: {},
             },
         },
+
+        // Clear the existing route so that we don't show an old route
+        routes: {
+            route0: {
+                geometry: {
+                    coordinates: null,
+                },
+            },
+        },
     });
 }
 
@@ -61,6 +71,19 @@ function saveWaypoint(transactionID, index, waypoint) {
         comment: {
             waypoints: {
                 [`waypoint${index}`]: waypoint,
+            },
+        },
+        // Empty out errors when we're saving a new waypoint as this indicates the user is updating their input
+        errorFields: {
+            route: null,
+        },
+
+        // Clear the existing route so that we don't show an old route
+        routes: {
+            route0: {
+                geometry: {
+                    coordinates: null,
+                },
             },
         },
     });
@@ -96,8 +119,72 @@ function removeWaypoint(transactionID, currentIndex) {
             ...transaction.comment,
             waypoints: reIndexedWaypoints,
         },
+        // Clear the existing route so that we don't show an old route
+        routes: {
+            route0: {
+                geometry: {
+                    coordinates: null,
+                },
+            },
+        },
     };
     Onyx.set(`${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`, newTransaction);
 }
 
-export {addStop, createInitialWaypoints, saveWaypoint, removeWaypoint};
+/**
+ * Gets the route for a set of waypoints
+ * Used so we can generate a map view of the provided waypoints
+ * @param {String} transactionID
+ * @param {Object} waypoints
+ */
+function getRoute(transactionID, waypoints) {
+    API.read(
+        'GetRoute',
+        {
+            transactionID,
+            waypoints: JSON.stringify(waypoints),
+        },
+        {
+            optimisticData: [
+                {
+                    // Clears any potentially stale error messages from fetching the route
+                    onyxMethod: Onyx.METHOD.MERGE,
+                    key: `${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`,
+                    value: {
+                        comment: {
+                            isLoading: true,
+                        },
+                        errorFields: {
+                            route: null,
+                        },
+                    },
+                },
+            ],
+            // The route and failure are sent back via pusher in the BE, we are just clearing the loading state here
+            successData: [
+                {
+                    onyxMethod: Onyx.METHOD.MERGE,
+                    key: `${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`,
+                    value: {
+                        comment: {
+                            isLoading: false,
+                        },
+                    },
+                },
+            ],
+            failureData: [
+                {
+                    onyxMethod: Onyx.METHOD.MERGE,
+                    key: `${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`,
+                    value: {
+                        comment: {
+                            isLoading: false,
+                        },
+                    },
+                },
+            ],
+        },
+    );
+}
+
+export {addStop, createInitialWaypoints, saveWaypoint, removeWaypoint, getRoute};
